Em/react ts glue 4 (#86)
* remove (unused) direct route to /preferences

* fixes reading list items not auto-deleting

* removes debounce on draft save

since drafts now only save onBlur, there's a much smaller risk of spamming draft saves, so it can be instant following a field being exited

* fixes draft removal

first, this removes the draft whether isMounted() or not. second, it dispatched fetchUser() so when we get back to homepage, the draft is already gone

* protects /form without login

* wraps PAWForm props in useCallback to avoid infinite render

* ppw can no longer be infinite; totalWeeks has a minimum of 1

* enforces that dates are available on Paper objects

* more concise syntax for date-picker

* updates test for 1 ppw to indicate the Infinity/week is not desired

* differentiates {User, Paper, Review}Response from pure type and casts incoming objects to front-end compliant versions

* gets rid of stale lastLogin field that is never used

* remove (unused) direct route to /preferences

* fixes reading list items not auto-deleting

* removes debounce on draft save

since drafts now only save onBlur, there's a much smaller risk of spamming draft saves, so it can be instant following a field being exited

* fixes draft removal

first, this removes the draft whether isMounted() or not. second, it dispatched fetchUser() so when we get back to homepage, the draft is already gone

* protects /form without login

* wraps PAWForm props in useCallback to avoid infinite render

* ppw can no longer be infinite; totalWeeks has a minimum of 1

* enforces that dates are available on Paper objects

* more concise syntax for date-picker

* updates test for 1 ppw to indicate the Infinity/week is not desired

* differentiates {User, Paper, Review}Response from pure type and casts incoming objects to front-end compliant versions

* gets rid of stale lastLogin field that is never used